requirements/dev.txt
cffi==1.14.1
click==7.1.2
Flask==1.1.2
Flask-Markdown==0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-SSE==0.2.1
flask-talisman==0.7.0
gevent==20.6.2
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.14.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
waitress==1.4.4
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0